Decouple the kubeseal CLI from the kubeseal library
